+++ cmd/cmd_helpers.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/fabmob/playground-standard-covoiturage/cmd/test/endpoint"
+	"github.com/fabmob/playground-standard-covoiturage/cmd/endpoint"
 	"github.com/spf13/cobra"
 	"github.com/stoewer/go-strcase"
 )
