+++ cmd/get_driver_journeys.go
@@ -4,8 +4,8 @@ import (
 	"net/http"
 	"net/url"
 
+	"github.com/fabmob/playground-standard-covoiturage/cmd/endpoint"
 	"github.com/fabmob/playground-standard-covoiturage/cmd/test"
-	"github.com/fabmob/playground-standard-covoiturage/cmd/test/endpoint"
 	"github.com/spf13/cobra"
 )
