+++ cmd/test/test_internals.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/fabmob/playground-standard-covoiturage/cmd/api"
-	"github.com/fabmob/playground-standard-covoiturage/cmd/test/endpoint"
+	"github.com/fabmob/playground-standard-covoiturage/cmd/endpoint"
 )
 
 // APIClient is a client to the API standard covoiturage
